added php package configuration

+++ build/vagrant/chef-repo/cookbooks-custom/base/recipes/default.rb
@@ -10,17 +10,29 @@ include_recipe "apt"
 execute "apt-mark-hold" do
   command "apt-mark hold grub-pc grub-pc-bin grub2-common && apt-get update"
   action :nothing
+  ignore_failure true
+  only_if { apt_installed? }
 end
 
 execute "apt-upgrade" do
   command "apt-get update && apt-get upgrade -y"
   action :nothing
+  ignore_failure true
+  only_if { apt_installed? }
   only_if { node['config']['drupical']['os_update'] }
 end
 
 execute "apt-cleanup" do
   command "apt-get -y autoremove && apt-get -y clean"
   action :nothing
+  ignore_failure true
+  only_if { apt_installed? }
+end
+
+template "/etc/apt/apt.conf.d/80dpkg" do
+  source "80dpkg"
+  mode '0666'
+  notifies :run, "execute[apt-get update]", :immediately
 end
 
 template "/etc/apt/apt.conf.d/90forceyes" do
